number_plate/requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
pytesseract==0.3.10